% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{generateParams}
\alias{generateParams}
\title{Generate matrix with all combinations of params ranges}
\usage{
generateParams(conf)
}
\arguments{
\item{conf}{minimum two elements named list of named vectors with ranges specification i.e. 'step', 'min', 'max'}
}
\value{
matrix with all combinations
}
\description{
Generate matrix with all combinations of params ranges
}
\author{
Konrad J. Debski
}
