% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{getCombinations}
\alias{getCombinations}
\title{Get combination of two vectors a and b}
\usage{
getCombinations(a, b)
}
\arguments{
\item{a}{vector.}

\item{b}{vector.}
}
\value{
list with replicated vector a and vector b
}
\description{
Get combination of two vectors a and b
}
\author{
Konrad J. Debski
}
