% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{getMultiCombinations}
\alias{getMultiCombinations}
\title{Get combination for list of vectors}
\usage{
getMultiCombinations(x)
}
\arguments{
\item{x}{list of vectors}
}
\value{
list with replicated vectors
}
\description{
Get combination for list of vectors
}
\author{
Konrad J. Debski
}
