% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{prepareParamsFromRanges}
\alias{prepareParamsFromRanges}
\title{Prepare params from ranges}
\usage{
prepareParamsFromRanges(conf)
}
\arguments{
\item{conf}{minimum two elements named list of named vectors with ranges specification i.e. 'step', 'min', 'max'}
}
\value{
named list with elements from ranges
}
\description{
Prepare params from ranges
}
\author{
Konrad J. Debski
}
